Refactor App Components Structure
