--- src/components/interfaces/index.js
@@ -1,4 +1,0 @@
-import TodayTasksInterface from "./TodayTasksInterface";
-import NewTaskInterface from "./NewTaskInterface";
-
-export { TodayTasksInterface, NewTaskInterface };
