+++ src/core-engine/checkToken.js
